Can detect winning moves
If you got 2 in a row/column/diagonal and the winning spot is open, the game will let you know there's an opening for you. It will also let you know if the spot you want is taken. (TBH I shouldn't have coded that but it helped with the detection)