Early menu UI
This menu UI offers a reset game option and a quit game option that are just placeholders. It also displays the score right in the middle. Another added line of text is "Click to continue" as we will need to make it so the user clicks to start playing, but the bot will go first. Last is another two liens of text in the pre-game screen to display how to open or close the menu.