Made X go first BUT game breaking bug
X can now go first regardless of player / bot, but there is a huge bug in there that will not allow the player to make a final move if all other spots are taken. The game will just hang and not allow the player to click. I've also noticed that the game sort of delays to restart the game.